refactor: RequestFactory 생성
